Remove actions - Remove UserRole
