--- auth/action/__init__.py
@@ -1,3 +1,0 @@
-# coding: utf-8
-from auth.action.role import RoleAction
-from auth.action.user_role import UserRoleAction
